# PERF: Fix quantile perf regression