Closes https://github.com/pandas-dev/pandas/issues/35049

```pyhon
591 µs ± 29.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

440 µs ± 22.8 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```