@jbrockmendel 2020-07-02
Should we take a more general look at where we can/should do string formatting more lazily?