@WillAyd 2020-07-02
Interesting - so the evaluation of q / 100 makes things slower right? I would have thought generally f-strings would be more performant given things like this:

https://stackoverflow.com/questions/43123408/f-strings-in-python-3-6

But I suppose that assumes that the expression needs to be evaluated, which isn't the case here unless an exception occurs